requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
marshmallow==3.5.0
pytz==2019.3
sqlparse==0.3.0
webargs==5.5.3
